Move default component to own file
